python/requirements.txt
python-liquid==2.0.2